remark: refactoring error handling
